chore: move legacy code
