+++ examples/pipeline/callbacks/main.go
@@ -3,8 +3,8 @@ package main
 import (
 	"context"
 
+	"github.com/henomis/lingoose/legacy/pipeline"
 	"github.com/henomis/lingoose/llm/openai"
-	"github.com/henomis/lingoose/pipeline"
 	"github.com/henomis/lingoose/prompt"
 	"github.com/henomis/lingoose/types"
 )
