+++ examples/pipeline/chat/main.go
@@ -5,11 +5,11 @@ import (
 	"encoding/json"
 	"fmt"
 
-	"github.com/henomis/lingoose/chat"
-	"github.com/henomis/lingoose/decoder"
+	"github.com/henomis/lingoose/legacy/chat"
+	"github.com/henomis/lingoose/legacy/decoder"
+	"github.com/henomis/lingoose/legacy/memory/ram"
+	"github.com/henomis/lingoose/legacy/pipeline"
 	"github.com/henomis/lingoose/llm/openai"
-	"github.com/henomis/lingoose/memory/ram"
-	"github.com/henomis/lingoose/pipeline"
 	"github.com/henomis/lingoose/prompt"
 	"github.com/henomis/lingoose/types"
 )
