+++ examples/pipeline/memory/main.go
@@ -5,10 +5,10 @@ import (
 	"encoding/json"
 	"fmt"
 
-	"github.com/henomis/lingoose/decoder"
+	"github.com/henomis/lingoose/legacy/decoder"
+	"github.com/henomis/lingoose/legacy/memory/ram"
+	"github.com/henomis/lingoose/legacy/pipeline"
 	llmmock "github.com/henomis/lingoose/llm/mock"
-	"github.com/henomis/lingoose/memory/ram"
-	"github.com/henomis/lingoose/pipeline"
 	"github.com/henomis/lingoose/prompt"
 )
