+++ examples/pipeline/simple/main.go
@@ -4,9 +4,9 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/henomis/lingoose/decoder"
+	"github.com/henomis/lingoose/legacy/decoder"
+	"github.com/henomis/lingoose/legacy/pipeline"
 	llmmock "github.com/henomis/lingoose/llm/mock"
-	"github.com/henomis/lingoose/pipeline"
 	"github.com/henomis/lingoose/prompt"
 )
