+++ examples/pipeline/splitter/main.go
@@ -5,9 +5,9 @@ import (
 	"encoding/json"
 	"fmt"
 
+	"github.com/henomis/lingoose/legacy/memory/ram"
+	"github.com/henomis/lingoose/legacy/pipeline"
 	"github.com/henomis/lingoose/llm/openai"
-	"github.com/henomis/lingoose/memory/ram"
-	"github.com/henomis/lingoose/pipeline"
 	"github.com/henomis/lingoose/prompt"
 	"github.com/henomis/lingoose/types"
 )
