+++ legacy/pipeline/decoder.go
