+++ legacy/pipeline/qa/qa.go
@@ -4,11 +4,11 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/henomis/lingoose/chat"
 	"github.com/henomis/lingoose/document"
 	"github.com/henomis/lingoose/index"
 	indexoption "github.com/henomis/lingoose/index/option"
-	"github.com/henomis/lingoose/pipeline"
+	"github.com/henomis/lingoose/legacy/chat"
+	"github.com/henomis/lingoose/legacy/pipeline"
 	"github.com/henomis/lingoose/prompt"
 	"github.com/henomis/lingoose/types"
 )
