+++ legacy/pipeline/splitter.go
