+++ legacy/pipeline/sql/sql.go
@@ -6,8 +6,8 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/henomis/lingoose/decoder"
-	"github.com/henomis/lingoose/pipeline"
+	"github.com/henomis/lingoose/legacy/decoder"
+	"github.com/henomis/lingoose/legacy/pipeline"
 	"github.com/henomis/lingoose/prompt"
 	"github.com/henomis/lingoose/types"
 )
