+++ legacy/pipeline/summarize/summarize.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 
 	"github.com/henomis/lingoose/document"
-	"github.com/henomis/lingoose/pipeline"
+	"github.com/henomis/lingoose/legacy/pipeline"
 	"github.com/henomis/lingoose/prompt"
 	"github.com/henomis/lingoose/types"
 )
