+++ llm/openai/legacy.go
@@ -9,7 +9,7 @@ import (
 	"os"
 	"strings"
 
-	"github.com/henomis/lingoose/chat"
+	"github.com/henomis/lingoose/legacy/chat"
 	"github.com/henomis/lingoose/llm/cache"
 	"github.com/henomis/lingoose/types"
 	"github.com/mitchellh/mapstructure"
